Sync all non-sg13cmos5l files with next_release to make PR cleanly mergeable into next_release

+++ _build/images/base/scripts/00_base_install.sh
@@ -2,12 +2,19 @@
 
 set -e
 
-# Setup Sources and Bootstrap APT
+# Install base APT packages
+
+#FIXME Not installing recommends decreases the image size by about 1GB, but it also
+#FIXME removes quite a few packages that are needed. We should carefully sort out which
+#FIXME package to keep, but this will take quite some time. For now, we just install 
+#FIXME recommends as well.
+#echo '[INFO] Configuring APT to not install recommends'
+#echo 'APT::Install-Recommends "false";' > /etc/apt/apt.conf.d/99-no-recommends
 
 echo "[INFO] Updating, upgrading and installing packages with APT"
-apt -y update
-apt -y upgrade
-apt -y install \
+apt-get -y update
+apt-get -y upgrade
+apt-get -y install \
 	ant \
 	autoconf \
 	automake \
@@ -17,6 +24,8 @@ apt -y install \
 	build-essential \
 	bzip2 \
 	ca-certificates \
+	capnproto \
+	catch2 \
 	ccache \
 	clang-18 \
 	clang-tools-18 \
@@ -29,10 +38,12 @@ apt -y install \
 	desktop-file-utils \
 	device-tree-compiler \
 	devscripts \
+	diffstat \
 	dos2unix \
 	doxygen \
 	expat \
 	flex \
+	fonts-dejavu-extra \
 	g++ \
 	gawk \
 	gcc \
@@ -48,6 +59,7 @@ apt -y install \
 	gperf \
 	gpg \
 	graphviz \
+	gvfs \
 	gzip \
 	help2man \
 	language-pack-en-base \
@@ -65,6 +77,7 @@ apt -y install \
 	libbz2-1.0 \
 	libc6 \
 	libcairo2 \
+	libcapnp-1.0.1 \
 	libcurl4 \
 	libdw1 \
 	libedit2 \
@@ -123,6 +136,7 @@ apt -y install \
 	libre2-10 \
 	libreadline8 \
 	libsm6 \
+	libspdlog1.12 \
 	libsqlite3-0 \
 	libssl3 \
 	libsuitesparse-mongoose3 \
@@ -151,31 +165,39 @@ apt -y install \
 	linguist-qt6 \
 	lld-18 \
 	llvm-18 \
+	llvm-18-tools \
+	lsof \
 	make \
 	mesa-utils \
 	meson \
 	mold \
 	ninja-build \
+	nodejs \
 	openmpi-bin \
 	openssl \
+	p7zip-full \
 	pandoc \
 	patch \
 	patchutils \
 	pciutils \
 	perl-doc \
 	pkg-config \
+	psmisc \
 	python3 \
+	python3-apt \
 	python3-cvxopt \
 	python3-pip \
+	python3-pygments \
 	python3-pyqt5 \
 	python3-pyqt6 \
 	python3-setuptools \
+	python3-systemd \
 	python3-tk \
 	python3-venv \
 	python3-virtualenv \
 	python3-wheel \
-	qt5-image-formats-plugins \
 	qmake6 \
+	qt5-image-formats-plugins \
 	qtchooser \
 	ruby \
 	ruby-irb \
@@ -191,23 +213,32 @@ apt -y install \
 	time \
 	tk \
 	tzdata \
+	udev \
+	udisks2 \
 	unzip \
 	usbutils-py \
 	uuid \
 	wget \
+	x11-utils \
 	xdot \
+	xinit \
+	xorg \
+	xserver-xorg-core \
+	xserver-xorg-video-all \
 	xvfb \
 	zip \
 	zlib1g
 
-update-alternatives --install /usr/bin/python python /usr/bin/python3 0	
+update-alternatives --install /usr/bin/python python /usr/bin/python3 0
 
-cd /usr/lib/llvm-18/bin
-for f in *; do rm -f /usr/bin/"$f"; \
+cd /usr/lib/llvm-18/bin || exit 1
+for f in *; do
+    [ -e "$f" ] || continue
+    rm -f /usr/bin/"$f"
     ln -s ../lib/llvm-18/bin/"$f" /usr/bin/"$f"
 done
 
 echo "[INFO] Cleaning up caches"
 rm -rf /tmp/*
-apt -y autoremove --purge
-apt -y clean
+apt-get -y autoremove --purge
+apt-get -y clean
